requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
requests==2.32.3
reportlab==4.2.2
google-generativeai==0.8.3
python-multipart==0.0.9
jinja2==3.1.4
aiofiles==24.1.0
